ShowConfig: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@40483 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/tools/ShowConfig.c
@@ -149,7 +149,7 @@ static VOID PrintProcessorInformation()
         }       
 
 
-        printf("PROCESSOR %d:\t[%s/%s] %s", i + 1, architecturestring, endiannessstring, modelstring);
+        printf("PROCESSOR %d:\t[%s/%s] %s", (int)(i + 1), architecturestring, endiannessstring, modelstring);
         if (cpuspeed)
             printf(" (%llu Mhz)", cpuspeed / 1000000);
         printf("\n");
